+++ pkg/codegen/go/gen_program_expressions.go
@@ -66,12 +66,12 @@ func (g *generator) GenAnonymousFunctionExpression(w io.Writer, expr *model.Anon
 	leadingSep := ""
 	for _, param := range expr.Signature.Parameters {
 		isInput := isInputty(param.Type)
-		g.Fgenf(w, "%s%s %s", leadingSep, param.Name, argumentTypeName(nil, param.Type, isInput))
+		g.Fgenf(w, "%s%s %s", leadingSep, param.Name, g.argumentTypeName(nil, param.Type, isInput))
 		leadingSep = ", "
 	}
 
 	isInput := isInputty(expr.Signature.ReturnType)
-	retType := argumentTypeName(nil, expr.Signature.ReturnType, isInput)
+	retType := g.argumentTypeName(nil, expr.Signature.ReturnType, isInput)
 	g.Fgenf(w, ") (%s, error) {\n", retType)
 
 	body, temps := g.lowerExpression(expr.Body, expr.Signature.ReturnType, isInput)
@@ -144,8 +144,10 @@ func (g *generator) GenFunctionCallExpression(w io.Writer, expr *model.FunctionC
 			g.genRelativeTraversalExpression(w, arg, isInput)
 		case *model.ScopeTraversalExpression:
 			g.genScopeTraversalExpression(w, arg, isInput)
+		case *model.ObjectConsExpression:
+			g.genObjectConsExpression(w, arg, expr.Type(), isInput)
 		default:
-			argType := argumentTypeName(arg, arg.Type(), isInput)
+			argType := g.argumentTypeName(arg, arg.Type(), isInput)
 			g.Fgenf(w, "%s(%v", argType, arg)
 			g.Fgenf(w, ")")
 		}
@@ -154,7 +156,8 @@ func (g *generator) GenFunctionCallExpression(w io.Writer, expr *model.FunctionC
 		case *model.TupleConsExpression:
 			g.genTupleConsExpression(w, arg, expr.Type())
 		case *model.ObjectConsExpression:
-			g.genObjectConsExpression(w, arg, expr.Type())
+			isInput := false
+			g.genObjectConsExpression(w, arg, expr.Type(), isInput)
 		case *model.LiteralValueExpression:
 			g.genLiteralValueExpression(w, arg, expr.Type())
 		default:
@@ -188,7 +191,7 @@ func (g *generator) GenFunctionCallExpression(w io.Writer, expr *model.FunctionC
 	case "fileAsset":
 		g.Fgenf(w, "pulumi.NewFileAsset(%.v)", expr.Args[0])
 	case hcl2.Invoke:
-		pkg, module, fn, diags := functionName(expr.Args[0])
+		pkg, module, fn, diags := g.functionName(expr.Args[0])
 		contract.Assert(len(diags) == 0)
 		if module == "" {
 			module = pkg
@@ -240,7 +243,7 @@ func (g *generator) GenLiteralValueExpression(w io.Writer, expr *model.LiteralVa
 }
 
 func (g *generator) genLiteralValueExpression(w io.Writer, expr *model.LiteralValueExpression, destType model.Type) {
-	argTypeName := argumentTypeName(expr, destType, false)
+	argTypeName := g.argumentTypeName(expr, destType, false)
 	isPulumiType := strings.HasPrefix(argTypeName, "pulumi.")
 
 	switch destType := destType.(type) {
@@ -298,19 +301,25 @@ func (g *generator) genLiteralValueExpression(w io.Writer, expr *model.LiteralVa
 }
 
 func (g *generator) GenObjectConsExpression(w io.Writer, expr *model.ObjectConsExpression) {
-	g.genObjectConsExpression(w, expr, expr.Type())
+	isInput := false
+	g.genObjectConsExpression(w, expr, expr.Type(), isInput)
 }
 
-func (g *generator) genObjectConsExpression(w io.Writer, expr *model.ObjectConsExpression, destType model.Type) {
+func (g *generator) genObjectConsExpression(
+	w io.Writer,
+	expr *model.ObjectConsExpression,
+	destType model.Type,
+	isInput bool,
+) {
 	if len(expr.Items) > 0 {
 		var temps []interface{}
-		isInput := isInputty(destType)
-		typeName := argumentTypeName(expr, destType, isInput)
+		isInput = isInput || isInputty(destType)
+		typeName := g.argumentTypeName(expr, destType, isInput)
 		if strings.HasSuffix(typeName, "Args") {
 			isInput = true
 		}
 		// invokes are not inputty
-		if strings.Contains(typeName, ".Lookup") {
+		if strings.Contains(typeName, ".Lookup") || strings.Contains(typeName, ".Get") {
 			isInput = false
 		}
 		isMap := strings.HasPrefix(typeName, "map[")
@@ -330,7 +339,7 @@ func (g *generator) genObjectConsExpression(w io.Writer, expr *model.ObjectConsE
 		}
 		g.genTemps(w, temps)
 
-		if isMap {
+		if isMap || !strings.HasSuffix(typeName, "Args") {
 			g.Fgenf(w, "%s", typeName)
 		} else {
 			g.Fgenf(w, "&%s", typeName)
@@ -339,7 +348,7 @@ func (g *generator) genObjectConsExpression(w io.Writer, expr *model.ObjectConsE
 
 		for _, item := range expr.Items {
 			if lit, ok := g.literalKey(item.Key); ok {
-				if isMap {
+				if isMap || strings.HasSuffix(typeName, "Map") {
 					g.Fgenf(w, "\"%s\"", lit)
 				} else {
 					g.Fgenf(w, "%s", Title(lit))
@@ -366,7 +375,7 @@ func (g *generator) genRelativeTraversalExpression(w io.Writer, expr *model.Rela
 		isInput = false
 	}
 	if isInput {
-		g.Fgenf(w, "%s(", argumentTypeName(expr, expr.Type(), isInput))
+		g.Fgenf(w, "%s(", g.argumentTypeName(expr, expr.Type(), isInput))
 	}
 	g.GenRelativeTraversalExpression(w, expr)
 	if isInput {
@@ -415,7 +424,7 @@ func (g *generator) genScopeTraversalExpression(w io.Writer, expr *model.ScopeTr
 
 	// TODO if it's an array type, we need a lowering step to turn []string -> pulumi.StringArray
 	if isInput {
-		g.Fgenf(w, "%s(", argumentTypeName(expr, expr.Type(), isInput))
+		g.Fgenf(w, "%s(", g.argumentTypeName(expr, expr.Type(), isInput))
 	}
 
 	// TODO: this isn't exhaustively correct as "range" could be a legit var name
@@ -488,7 +497,7 @@ func (g *generator) genTupleConsExpression(w io.Writer, expr *model.TupleConsExp
 		expr.Expressions[i] = item
 	}
 	g.genTemps(w, temps)
-	argType := argumentTypeName(expr, destType, isInput)
+	argType := g.argumentTypeName(expr, destType, isInput)
 	g.Fgenf(w, "%s{\n", argType)
 	switch len(expr.Expressions) {
 	case 0:
@@ -514,7 +523,7 @@ func (g *generator) GenUnaryOpExpression(w io.Writer, expr *model.UnaryOpExpress
 }
 
 // argumentTypeName computes the go type for the given expression and model type.
-func argumentTypeName(expr model.Expression, destType model.Type, isInput bool) string {
+func (g *generator) argumentTypeName(expr model.Expression, destType model.Type, isInput bool) string {
 	var tokenRange hcl.Range
 	if expr != nil {
 		node := expr.SyntaxNode()
@@ -526,22 +535,39 @@ func argumentTypeName(expr model.Expression, destType model.Type, isInput bool) 
 		switch schemaType := schemaType.(type) {
 		case *schema.ArrayType:
 			token := schemaType.ElementType.(*schema.ObjectType).Token
-			_, module, member, diags := hcl2.DecomposeToken(token, tokenRange)
+			pkg, module, member, diags := hcl2.DecomposeToken(token, tokenRange)
+			// namespaceless invokes
+			if module == "" || strings.HasPrefix(module, "/") || strings.HasPrefix(module, "index/") {
+				module = pkg
+			}
 			importPrefix := strings.Split(module, "/")[0]
 			contract.Assert(len(diags) == 0)
 			fmtString := "[]%s.%s"
 			if isInput {
+				member = Title(member)
+				if strings.HasPrefix(member, "Get") {
+					if g.useLookupInvokeForm(token) {
+						member = strings.Replace(member, "Get", "Lookup", 1)
+					}
+					return fmt.Sprintf("[]%s.%s", importPrefix, member)
+				}
 				fmtString = "%s.%sArray"
 			}
 			return fmt.Sprintf(fmtString, importPrefix, member)
 		case *schema.ObjectType:
 			token := schemaType.Token
-			_, module, member, diags := hcl2.DecomposeToken(token, tokenRange)
+			pkg, module, member, diags := hcl2.DecomposeToken(token, tokenRange)
+			// namespaceless invokes
+			if module == "" || strings.HasPrefix(module, "/") || strings.HasPrefix(module, "index/") {
+				module = pkg
+			}
 			importPrefix := strings.Split(module, "/")[0]
 			contract.Assert(len(diags) == 0)
 			member = Title(member)
 			if strings.HasPrefix(member, "Get") {
-				member = strings.Replace(member, "Get", "Lookup", 1)
+				if g.useLookupInvokeForm(token) {
+					member = strings.Replace(member, "Get", "Lookup", 1)
+				}
 				return fmt.Sprintf("%s.%s", importPrefix, member)
 			}
 			fmtString := "%s.%s"
@@ -586,13 +612,13 @@ func argumentTypeName(expr model.Expression, destType model.Type, isInput bool) 
 		}
 		return "map[string]interface{}"
 	case *model.MapType:
-		valType := argumentTypeName(nil, destType.ElementType, isInput)
+		valType := g.argumentTypeName(nil, destType.ElementType, isInput)
 		if isInput {
 			return fmt.Sprintf("pulumi.%sMap", Title(valType))
 		}
 		return fmt.Sprintf("map[string]%s", valType)
 	case *model.ListType:
-		argTypeName := argumentTypeName(nil, destType.ElementType, isInput)
+		argTypeName := g.argumentTypeName(nil, destType.ElementType, isInput)
 		if strings.HasPrefix(argTypeName, "pulumi.") {
 			return fmt.Sprintf("%sArray", argTypeName)
 		}
@@ -613,7 +639,7 @@ func argumentTypeName(expr model.Expression, destType model.Type, isInput bool) 
 		}
 
 		if elmType != nil {
-			argTypeName := argumentTypeName(nil, elmType, isInput)
+			argTypeName := g.argumentTypeName(nil, elmType, isInput)
 			if strings.HasPrefix(argTypeName, "pulumi.") {
 				return fmt.Sprintf("%sArray", argTypeName)
 			}
@@ -626,16 +652,16 @@ func argumentTypeName(expr model.Expression, destType model.Type, isInput bool) 
 		return "[]interface{}"
 	case *model.OutputType:
 		isInput = true
-		return argumentTypeName(expr, destType.ElementType, isInput)
+		return g.argumentTypeName(expr, destType.ElementType, isInput)
 	case *model.UnionType:
 		for _, ut := range destType.ElementTypes {
 			if _, isOpaqueType := ut.(*model.OpaqueType); isOpaqueType {
-				return argumentTypeName(expr, ut, isInput)
+				return g.argumentTypeName(expr, ut, isInput)
 			}
 		}
 		return "interface{}"
 	case *model.PromiseType:
-		return argumentTypeName(expr, destType.ElementType, isInput)
+		return g.argumentTypeName(expr, destType.ElementType, isInput)
 	default:
 		contract.Failf("unexpected destType type %T", destType)
 	}
@@ -737,7 +763,7 @@ func (g *generator) genApply(w io.Writer, expr *model.FunctionCallExpression) {
 	applyArgs, then := hcl2.ParseApplyCall(expr)
 	then = stripInputs(then).(*model.AnonymousFunctionExpression)
 	isInput := false
-	retType := argumentTypeName(nil, then.Signature.ReturnType, isInput)
+	retType := g.argumentTypeName(nil, then.Signature.ReturnType, isInput)
 	// TODO account for outputs in other namespaces like aws
 	typeAssertion := fmt.Sprintf(".(%sOutput)", retType)
 	if !strings.HasPrefix(retType, "pulumi.") {
@@ -755,23 +781,14 @@ func (g *generator) genApply(w io.Writer, expr *model.FunctionCallExpression) {
 		}
 		// TODO need lowering step to rewrite then argument references
 		// in terms of scope traversal from all result: val[i] etc.
-		g.Fgenf(w, ").Apply(%.v)%s", then, typeAssertion)
+		g.Fgenf(w, ").ApplyT(%.v)%s", then, typeAssertion)
 	}
 }
 
 func (g *generator) genStringLiteral(w io.Writer, v string) {
-	// TODO more robust and go-specific handling of strings
-	newlines := strings.Contains(v, "\n")
-	if !newlines {
-		// This string does not contain newlines so we'll generate a regular string literal. Quotes and backslashes
-		// will be escaped in conformance with
-		// https://golang.org/ref/spec#String_literals
-		g.Fgen(w, "\"")
-		g.Fgen(w, g.escapeString(v))
-		g.Fgen(w, "\"")
-	} else {
-		g.genNYI(w, "TODO multiline strings")
-	}
+	g.Fgen(w, "\"")
+	g.Fgen(w, g.escapeString(v))
+	g.Fgen(w, "\"")
 }
 
 func (g *generator) escapeString(v string) string {
@@ -779,6 +796,11 @@ func (g *generator) escapeString(v string) string {
 	for _, c := range v {
 		if c == '"' || c == '\\' {
 			builder.WriteRune('\\')
+		}
+		if c == '\n' {
+			builder.WriteRune('\\')
+			builder.WriteRune('n')
+			continue
 		}
 		builder.WriteRune(c)
 	}
@@ -831,17 +853,16 @@ func (g *generator) literalKey(x model.Expression) (string, bool) {
 }
 
 // functionName computes the go package, module, and name for the given function token.
-func functionName(tokenArg model.Expression) (string, string, string, hcl.Diagnostics) {
+func (g *generator) functionName(tokenArg model.Expression) (string, string, string, hcl.Diagnostics) {
 	token := tokenArg.(*model.TemplateExpression).Parts[0].(*model.LiteralValueExpression).Value.AsString()
 	tokenRange := tokenArg.SyntaxNode().Range()
 
 	// Compute the resource type from the Pulumi type token.
 	pkg, module, member, diagnostics := hcl2.DecomposeToken(token, tokenRange)
-	// GetResource -> LookupResource
-	// TODO figure out how to deal with get/lookup ambiguity
-	// https://github.com/pulumi/pulumi/blob/adfa8116fbf61ad328568fe4b53e32c751f30553/pkg/codegen/go/gen.go#L1171
 	if strings.HasPrefix(member, "get") {
-		member = strings.Replace(member, "get", "lookup", 1)
+		if g.useLookupInvokeForm(token) {
+			member = strings.Replace(member, "get", "lookup", 1)
+		}
 	}
 	return pkg, strings.Replace(module, "/", ".", -1), Title(member), diagnostics
 }